major restructure of App project, cleaned up some warnings
