+++ src/DateNight.App/Components/IdeaComponent/IdeaComponent.razor
