--- src/DateNight.App/Models/LoginModel.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace DateNight.App.Models;
-
-public class LoginModel
-{
-    [Required, EmailAddress]
-    public string Email { get; set; }
-
-    [Required]
-    public string Password { get; set; }
-}
