+++ src/DateNight.App/Pages/Account/AccountModel.cs
@@ -1,15 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace DateNight.App.Models;
+namespace DateNight.App.Pages.Account;
 
-internal class CreateAccountModel
+public class AccountModel
 {
     [Required, EmailAddress]
     public string Email { get; set; } = string.Empty;
 
     [Required]
     public string Name { get; set; } = string.Empty;
-
-    [Required, MinLength(8)]
-    public string Password { get; set; } = string.Empty;
 }
