+++ src/DateNight.App/Pages/ActiveDate/ActiveDate.razor
@@ -1,4 +1,5 @@
 ﻿@page "/ideas/active";
+@using DateNight.App.Components.IdeaComponent
 
 <div class="d-flex flex-column h-100">
     <h2>Active Date</h2>
