+++ src/DateNight.App/Pages/Login/LoginModel.cs
@@ -1,12 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace DateNight.App.Models;
+namespace DateNight.App.Pages.Login;
 
-public class AccountModel
+public class LoginModel
 {
     [Required, EmailAddress]
     public string Email { get; set; } = string.Empty;
 
     [Required]
-    public string Name { get; set; } = string.Empty;
+    public string Password { get; set; } = string.Empty;
 }
