+++ src/DateNight.App/Pages/RandomIdea/RandomIdea.razor
@@ -1,4 +1,5 @@
 ﻿@page "/ideas/random"
+@using DateNight.App.Components.IdeaComponent
 
 @if (_isFirstLoad)
 {
