+++ src/DateNight.App/Pages/ViewIdeas/ViewIdeas.razor
@@ -1,4 +1,5 @@
 ﻿@page "/ideas/view"
+@using DateNight.App.Components.IdeaComponent
 
 <h3>View Ideas</h3>
