+++ src/DateNight.App/Shared/NavMenu.razor
@@ -37,13 +37,4 @@
     </nav>
 </div>
 
-@code {
-    private bool collapseNavMenu = true;
 
-    private string NavMenuCssClass => collapseNavMenu ? "collapse" : null;
-
-    private void ToggleNavMenu()
-    {
-        collapseNavMenu = !collapseNavMenu;
-    }
-}
