+++ src/DateNight.App/Shared/NavMenu.razor.cs
@@ -1,0 +1,11 @@
+namespace DateNight.App.Shared;
+
+public partial class NavMenu
+{
+    private bool _collapseNavMenu = true;
+    private string NavMenuCssClass => _collapseNavMenu ? "collapse" : "";
+    private void ToggleNavMenu()
+    {
+        _collapseNavMenu = !_collapseNavMenu;
+    }
+}
